gmsearch rewritten
